15일차/pyproject.toml
[tool.poetry]
name = "django-restaurant-review"
version = "0.1.0"
description = "Django restaurant review project"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
django = ">=5.2.5"
django-cleanup = ">=9.0.0"
django-extensions = ">=4.1"
django-restframework = ">=0.0.1"
djangorestframework = ">=3.16.1"
djangorestframework-simplejwt = ">=5.5.1"
ipython = ">=9.4.0"
mysqlclient = ">=2.2.7"
pillow = ">=11.3.0"
django-storages = ">=1.14.6,<2.0.0"
boto3 = ">=1.40.8,<2.0.0"

[tool.poetry.group.dev.dependencies]
gunicorn = "^23.0.0"

